Split the health check test into separate classes
