+++ src/test/java/endpoints/README.md
@@ -1,0 +1,9 @@
+# restful-booker endpoints
+
+The following endpoints are available in the restful-booker web API.
+
+- /auth
+- /booking
+- /ping
+                        
+Calls to anything else will return a 404 Not Found response.
